TRY: unify config in config.json, allow overwriting site.js

+++ src/bin/build.sh
@@ -9,7 +9,8 @@ fi
 ( cd site ; if ! [ -d node_modules ]; then npm ci --legacy-peer-deps ; fi )
 
 cp -R _static/. site/public/ #NodeJS no maneja mover y copiar archivos que no sean texto.
-cp site.js site/src/config/site.js
+if [ -f site.js ]; then cp site.js site/src/config/site.js ; fi
+if ! [ -f config.json ]; then echo "{}" > config.json ; fi
 
 node site/src/lib/to-next.mjs . ./site/src/app ./site/public
